refactor(section): 별도 패키지로 분리
