+++ src/main/java/subway/section/SectionDao.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.section;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
