+++ src/main/java/subway/section/Sections.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.section;
 
 import java.util.Collections;
 import java.util.Comparator;
